pyproject.toml
[project]
name = "ppt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "comtypes>=1.4.8",
    "pillow>=11.0.0",
    "python-pptx>=1.0.2",
    "pywin32>=308",
]
